requirements.txt
google-generativeai==0.7.2
requests==2.31.0
